backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
jinja2==3.1.4
requests==2.32.3
pytest==8.3.2
httpx==0.27.0
pydantic-settings==2.4.0
google-generativeai==0.8.3
PyPDF2==3.0.1
Pillow==10.4.0
python-magic==0.4.27